# Implement redeem and redeemWith without attempt