Attempts #3148 

The benchmarks show mixed results with only the happy path being worse:
```
Before:
Benchmark                        (size)   Mode  Cnt     Score     Error  Units
RedeemWithBenchmark.errorRaised   10000  thrpt   10  2416.151 ±  56.584  ops/s
RedeemWithBenchmark.happyPath     10000  thrpt   10  3278.920 ± 157.717  ops/s
RedeemBenchmark.errorRaised       10000  thrpt   10  1432.012 ±  10.972  ops/s
RedeemBenchmark.happyPath         10000  thrpt   10  2099.513 ±   4.591  ops/s

After:
Benchmark                        (size)   Mode  Cnt     Score    Error  Units
RedeemWithBenchmark.errorRaised   10000  thrpt   20  3315.182 ±  16.192  ops/s
RedeemWithBenchmark.happyPath     10000  thrpt   20  4765.599 ±  91.358  ops/s
RedeemBenchmark.errorRaised       10000  thrpt   10  1639.106 ±  19.014  ops/s
RedeemBenchmark.happyPath         10000  thrpt   10  1785.519 ±  11.161  ops/s
```