@djspiewak 2022-10-26
This is really nice! Thank you! I'm a bit concerned about the happy path degradation though, since it suggests that the run loop itself could have gotten slower. Given the performance sensitivity of those particular tableswitches, I think it might be worth running the Map and Bind benchmarks, and perhaps also the scheduler benchmark, before and after. That should give us a decent idea of whether we've accidentally degraded other cases by specializing this one.