@JPonte 2022-10-29
Thanks for the feedback! I'm afraid you're right, seems like it has reduced the performance of the other cases ever so slightly... 
It was still fun to try it out 🥲 
```
Before:
MapCallsBenchmark.batch120                                         N/A      N/A  thrpt   20    525.883 ±    2.815    ops/s
MapCallsBenchmark.batch30                                          N/A      N/A  thrpt   20    137.577 ±    0.420    ops/s
MapCallsBenchmark.one                                              N/A      N/A  thrpt   20      4.729 ±    0.139    ops/s
DeepBindBenchmark.async                                            N/A    10000  thrpt   20   2306.285 ±   17.682    ops/s
DeepBindBenchmark.delay                                            N/A    10000  thrpt   20   7457.217 ±   26.099    ops/s
DeepBindBenchmark.pure                                             N/A    10000  thrpt   20   9350.599 ±   44.264    ops/s
ShallowBindBenchmark.async                                         N/A    10000  thrpt   20   1537.905 ±   61.159    ops/s
ShallowBindBenchmark.delay                                         N/A    10000  thrpt   20   7666.238 ±   86.941    ops/s
ShallowBindBenchmark.pure                                          N/A    10000  thrpt   20   8839.406 ±   19.938    ops/s
WorkStealingBenchmark.alloc                                        N/A  1000000  thrpt   20      8.412 ±    0.034  ops/min
WorkStealingBenchmark.manyThreadsSchedulingBenchmark               N/A  1000000  thrpt   20     24.517 ±    0.310  ops/min
WorkStealingBenchmark.runnableScheduling                           N/A  1000000  thrpt   20   2033.663 ±    3.173  ops/min
WorkStealingBenchmark.runnableSchedulingScalaGlobal                N/A  1000000  thrpt   20   1655.517 ±    2.734  ops/min
WorkStealingBenchmark.scheduling                                   N/A  1000000  thrpt   20     26.716 ±    0.725  ops/min


After:
MapCallsBenchmark.batch120                                         N/A      N/A  thrpt   20   522.967 ±     9.867    ops/s
MapCallsBenchmark.batch30                                          N/A      N/A  thrpt   20   138.529 ±     0.940    ops/s
MapCallsBenchmark.one                                              N/A      N/A  thrpt   20     4.732 ±     0.010    ops/s
DeepBindBenchmark.async                                            N/A    10000  thrpt   20  2253.959 ±    18.358    ops/s
DeepBindBenchmark.delay                                            N/A    10000  thrpt   20  7320.230 ±    17.911    ops/s
DeepBindBenchmark.pure                                             N/A    10000  thrpt   20  9223.594 ±    54.671    ops/s
ShallowBindBenchmark.async                                         N/A    10000  thrpt   20  1310.506 ±    75.333    ops/s
ShallowBindBenchmark.delay                                         N/A    10000  thrpt   20  7570.771 ±     6.065    ops/s
ShallowBindBenchmark.pure                                          N/A    10000  thrpt   20  8800.231 ±    49.698    ops/s
WorkStealingBenchmark.alloc                                        N/A  1000000  thrpt   20     8.362 ±     0.058  ops/min
WorkStealingBenchmark.manyThreadsSchedulingBenchmark               N/A  1000000  thrpt   20    24.968 ±     0.203  ops/min
WorkStealingBenchmark.runnableScheduling                           N/A  1000000  thrpt   20  1878.291 ±    87.140  ops/min
WorkStealingBenchmark.runnableSchedulingScalaGlobal                N/A  1000000  thrpt   20  1649.051 ±     7.617  ops/min
WorkStealingBenchmark.scheduling                                   N/A  1000000  thrpt   20    25.246 ±     1.082  ops/min
```