@djspiewak 2022-10-29
Yeah this is just *barely* outside the margin for error. Considering it's consistently vectored, I'd say that it probably is a very slight reduction in overall performance. That's really unfortunate. :-(